Structure | Gabriel Blanco

+++ src/main/java/co/edu/unbosque/model/UsuarioDAO.java
@@ -3,5 +3,5 @@ package co.edu.unbosque.model;
 import co.edu.unbosque.model.persistence.UsuarioDTO;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsuarioDAO extends CrudRepository<UsuarioDTO, Integer> {
+public interface UsuarioDAO extends CrudRepository<UsuarioDTO, Long> {
 }
